tasks/pyproject.toml
[tool.poetry]
name = "milestone"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.6.0"
async-timeout = "4.0.2"
attrs = "22.2.0"
certifi = "2022.12.7"
charset-normalizer = "3.0.1"
coreapi = "2.3.3"
coreschema = "0.0.4"
coverage = "7.2.0"
django = "4.1.6"
django-cors-headers = "3.13.0"
django-filter = "22.1"
django-nose = "1.4.7"
django-redis = "5.2.0"
django-rest-swagger = "2.2.0"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.2.2"
drf-util = "1.0.17"
drf-yasg = "1.21.5"
executing = "1.2.0"
idna = "3.4"
inflection = "0.5.1"
itypes = "1.2.0"
jinja2 = "3.1.2"
jsonschema = "4.17.3"
markupsafe = "2.1.2"
nose = "1.3.7"
openapi-codec = "1.3.2"
packaging = "23.0"
psycopg2-binary = "2.9.5"
pyjwt = "2.6.0"
pyrsistent = "0.19.3"
python-dateutil = "2.8.2"
pytz = "2022.7.1"
pyyaml = "6.0"
freezegun = "1.2.2"
redis = "4.5.1"
requests = "2.28.2"
ruamel-yaml = "0.17.21"
ruamel-yaml-clib = "0.2.7"
simplejson = "3.18.3"
six = "1.16.0"
sqlparse = "0.4.3"
swagger-spec-validator = "3.0.3"
typing-extensions = "4.5.0"
tzdata = "2022.7"
uritemplate = "4.1.1"
urllib3 = "1.26.14"
varname = "0.10.0"
django-extensions = "^3.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
